+++ src/main/java/com/jerry/mekaf/common/upgrade/ItemChemicalToChemicalUpgradeData.java
@@ -3,9 +3,9 @@ package com.jerry.mekaf.common.upgrade;
 import mekanism.api.chemical.IChemicalTank;
 import mekanism.api.energy.IEnergyContainer;
 import mekanism.api.inventory.IInventorySlot;
+import mekanism.common.inventory.slot.ChemicalInventorySlot;
 import mekanism.common.inventory.slot.EnergyInventorySlot;
 import mekanism.common.inventory.slot.InputInventorySlot;
-import mekanism.common.inventory.slot.chemical.ChemicalInventorySlot;
 import mekanism.common.tile.component.ITileComponent;
 import mekanism.common.tile.interfaces.IRedstoneControl;
 
@@ -19,17 +19,17 @@ public class ItemChemicalToChemicalUpgradeData extends ItemToChemicalUpgradeData
 
     public final ChemicalInventorySlot chemicalSlot;
     public final IChemicalTank inputTank;
-    public final long[] usedSoFar;
+    public final int[] usedSoFar;
 
     public ItemChemicalToChemicalUpgradeData(HolderLookup.Provider provider, boolean redstone, IRedstoneControl.RedstoneControl controlType,
-                                             IEnergyContainer energyContainer, int operatingTicks, long usedSoFar, EnergyInventorySlot energySlot,
+                                             IEnergyContainer energyContainer, int operatingTicks, int usedSoFar, EnergyInventorySlot energySlot,
                                              ChemicalInventorySlot chemicalSlot, InputInventorySlot inputSlot, IChemicalTank inputTank, IChemicalTank outputTank,
                                              List<ITileComponent> components, PathElement problemPath) {
-        this(provider, redstone, controlType, energyContainer, new int[] { operatingTicks }, new long[] { usedSoFar }, energySlot, chemicalSlot, Collections.singletonList(inputSlot), inputTank, Collections.singletonList(outputTank), false, components, problemPath);
+        this(provider, redstone, controlType, energyContainer, new int[] { operatingTicks }, new int[] { usedSoFar }, energySlot, chemicalSlot, Collections.singletonList(inputSlot), inputTank, Collections.singletonList(outputTank), false, components, problemPath);
     }
 
     public ItemChemicalToChemicalUpgradeData(HolderLookup.Provider provider, boolean redstone, IRedstoneControl.RedstoneControl controlType,
-                                             IEnergyContainer energyContainer, int[] progress, long[] usedSoFar, EnergyInventorySlot energySlot, ChemicalInventorySlot chemicalSlot,
+                                             IEnergyContainer energyContainer, int[] progress, int[] usedSoFar, EnergyInventorySlot energySlot, ChemicalInventorySlot chemicalSlot,
                                              List<IInventorySlot> inputSlots, IChemicalTank inputTank, List<IChemicalTank> outputTanks, boolean sorting, List<ITileComponent> components, PathElement problemPath) {
         super(provider, redstone, controlType, energyContainer, progress, energySlot, inputSlots, outputTanks, sorting, components, problemPath);
         this.chemicalSlot = chemicalSlot;
